Solved part 2
In this commit we add more data (the 'count'
for each child rule) so we can calculate
how many bags a certain rule contains.

We add an 'adjustment' flag of sorts to the recursive
calculator because at the bottom two
levels of our treeish structure, we don't
need to add any extra to solve, but for internal rules above the
bottom two levels, the fact that there is a child rule should add
one extra bag to the calculation.