Solved part 2 with some caching to deal with perf
Worked with Timothy Hoffman (@Waldo1637) to solve
the crux of the problem. Whenever the combination
branches, we can mentally model that as a tree
that is branching into sub trees. All subtrees
will end with the last element of the input,
and thus the solution to the problem is just the
number of leaves in this tree.

To calculate this, we can use recursion to 'branch'
and List.fold inside the recursion for each subtree.

Trying to do that naively worked well for the small
and medium inputs, but the main input took forever.
Adding a cache of the results for each step of the
iteration dramatically helped and solved it!